requirements.txt
pythonosc
paho-mqtt